moving human_readable_bytes to uv-console
