+++ crates/uv-resolver/src/lock/tree.rs
@@ -14,6 +14,7 @@ use uv_normalize::{ExtraName, GroupName, PackageName};
 use uv_pep440::Version;
 use uv_pep508::MarkerTree;
 use uv_pypi_types::ResolverMarkerEnvironment;
+use uv_console::human_readable_bytes;
 
 use crate::lock::PackageId;
 use crate::{Lock, PackageMap};
@@ -679,20 +680,4 @@ impl std::fmt::Display for TreeDisplay<'_> {
 
         Ok(())
     }
-}
-
-/// Formats a number of bytes into a human readable SI-prefixed size (binary units).
-///
-/// Returns a tuple of `(quantity, units)`.
-#[allow(
-    clippy::cast_possible_truncation,
-    clippy::cast_possible_wrap,
-    clippy::cast_precision_loss,
-    clippy::cast_sign_loss
-)]
-fn human_readable_bytes(bytes: u64) -> (f32, &'static str) {
-    const UNITS: [&str; 7] = ["B", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB"];
-    let bytes_f32 = bytes as f32;
-    let i = ((bytes_f32.log2() / 10.0) as usize).min(UNITS.len() - 1);
-    (bytes_f32 / 1024_f32.powi(i as i32), UNITS[i])
 }
